refactor: reduce code duplication between RollbarClient and RollbarBlazorClient implementations.
